BATTERY POWER: fixed bugged function that i added long time ago.
*We do not need to update the thermal_mitigation each battery update.
it's creating not needed work non stop on charge.
*When current_now is below 300ma then set thermal_mitigation to 300
in order to maintain stable connection to device over USB. no disconnects.
*If battery is LOW and ROM we connected strong charger, then each time ROM
Theraml update the smb349_thermal_mitigation check what we want to have based
on fast charge profile. and update it and not what ROM thermal want.
*Check if battery overheatted or 55C and drop thermal_mitigation to 300
to cool it and preserve.

Conflicts:
	drivers/power/smb349_charger.c